Fix indirect imports/exports
