+++ auto_gptq_next/__init__.py
@@ -1,3 +1,4 @@
 from .models import AutoGPTQNext, BaseQuantizeConfig, QuantizeConfig
+from .quantization import BaseQuantizeConfig, QuantizeConfig
 from .utils.exllama import exllama_set_max_input_length
 from .version import __version__
